fix: final RRYautja API error fixes for 1.6 (145->4 errors)

+++ Source/RRYautja/Harmony/Patches/AvP_Pawn_PathFollower_CostToMoveIntoCell_Patch.cs
@@ -26,11 +26,11 @@ namespace RRYautja
                 int num;
                 if (c.x == pawn.Position.x || c.z == pawn.Position.z)
                 {
-                    num = pawn.TicksPerMoveCardinal;
+                    num = (int)pawn.TicksPerMoveCardinal;
                 }
                 else
                 {
-                    num = pawn.TicksPerMoveDiagonal;
+                    num = (int)pawn.TicksPerMoveDiagonal;
                 }
 
                 Building edifice = c.GetEdifice(pawn.Map);
